moved workspace and its tests to gymlib
