+++ benchmark/tests/integtest_benchmark.py
@@ -2,10 +2,15 @@ import shutil
 import unittest
 from pathlib import Path
 
-from gymlib import (
+from gymlib.symlinks_paths import (
     get_tables_symlink_path,
     get_workload_suffix,
     get_workload_symlink_path,
+)
+from gymlib.workspace import (
+    DBGymWorkspace,
+    fully_resolve_path,
+    get_workspace_path_from_config,
 )
 
 # It's ok to import private functions from the benchmark module because this is an integration test.
@@ -13,11 +18,6 @@ from benchmark.constants import DEFAULT_SCALE_FACTOR
 from benchmark.job.cli import _job_tables, _job_workload
 from benchmark.tpch.cli import _tpch_tables, _tpch_workload
 from benchmark.tpch.constants import DEFAULT_TPCH_SEED
-from util.workspace import (
-    DBGymWorkspace,
-    fully_resolve_path,
-    get_workspace_path_from_config,
-)
 
 
 class BenchmarkTests(unittest.TestCase):
