+++ benchmark/tpch/load_info.py
@@ -1,10 +1,10 @@
 from pathlib import Path
 from typing import Optional
 
-from gymlib import get_tables_symlink_path
+from gymlib.symlinks_paths import get_tables_symlink_path
+from gymlib.workspace import DBGymWorkspace, fully_resolve_path
 
 from dbms.load_info_base_class import LoadInfoBaseClass
-from util.workspace import DBGymWorkspace, fully_resolve_path
 
 TPCH_SCHEMA_FNAME = "tpch_schema.sql"
 TPCH_CONSTRAINTS_FNAME = "tpch_constraints.sql"
