+++ dbms/postgres/cli.py
@@ -13,20 +13,14 @@ import pglast
 import psutil
 import psycopg
 import sqlalchemy
-from gymlib import (
+from gymlib.shell import subprocess_run
+from gymlib.symlinks_paths import (
     get_dbdata_tgz_symlink_path,
     get_pgbin_symlink_path,
     get_repo_symlink_path,
     linkname_to_name,
 )
-from gymlib.shell import subprocess_run
-from sqlalchemy import create_engine, text
-
-from benchmark.constants import DEFAULT_SCALE_FACTOR
-from benchmark.job.load_info import JobLoadInfo
-from benchmark.tpch.load_info import TpchLoadInfo
-from dbms.load_info_base_class import LoadInfoBaseClass
-from util.workspace import (
+from gymlib.workspace import (
     WORKSPACE_PATH_PLACEHOLDER,
     DBGymWorkspace,
     fully_resolve_path,
@@ -34,6 +28,12 @@ from util.workspace import (
     is_fully_resolved,
     is_ssd,
 )
+from sqlalchemy import create_engine, text
+
+from benchmark.constants import DEFAULT_SCALE_FACTOR
+from benchmark.job.load_info import JobLoadInfo
+from benchmark.tpch.load_info import TpchLoadInfo
+from dbms.load_info_base_class import LoadInfoBaseClass
 
 DBGYM_POSTGRES_USER = "dbgym_user"
 DBGYM_POSTGRES_PASS = "dbgym_pass"
