+++ gymlib_package/gymlib/__init__.py
@@ -1,5 +1,1 @@
-# Everything at the "core" of gymlib will be imported directly (e.g. `from .[module] import *`).
-# Everything not as important will be only be imported as a module (e.g. `from . import [module]`).
-from . import shell
-from .magic import *
-from .symlinks_paths import *
+from . import magic, shell, symlinks_paths, workspace
