+++ gymlib_package/gymlib/tests/unittest_filesystem_unittest_util.py
@@ -3,7 +3,7 @@ import shutil
 import unittest
 from pathlib import Path
 
-from util.tests.filesystem_unittest_util import (
+from gymlib.tests.filesystem_unittest_util import (
     FilesystemStructure,
     create_structure,
     verify_structure,
