+++ task.py
@@ -1,9 +1,9 @@
 import click
+from gymlib.workspace import make_standard_dbgym_workspace
 
 from benchmark.cli import benchmark_group
 from dbms.cli import dbms_group
 from manage.cli import manage_group
-from util.workspace import make_standard_dbgym_workspace
 
 # TODO(phw2): Save commit, git diff, and run command.
 # TODO(phw2): Remove write permissions on old run_*/ dirs to enforce that they are immutable.
